# Remove unnecessary work-around for HIP