--- a/src/details/ArborX_DetailsHappyTreeFriends.hpp
+++ b/src/details/ArborX_DetailsHappyTreeFriends.hpp
@@ -44,27 +44,14 @@ struct HappyTreeFriends
   }
 
   template <class BVH>
-  static KOKKOS_FUNCTION
-// FIXME_HIP See https://github.com/arborx/ArborX/issues/553
-#ifdef __HIP_DEVICE_COMPILE__
-      auto
-#else
-      auto const &
-#endif
-      getInternalBoundingVolume(BVH const &bvh, int i)
+  static KOKKOS_FUNCTION auto const &getInternalBoundingVolume(BVH const &bvh,
+                                                               int i)
   {
     return bvh._internal_nodes(internalIndex(bvh, i)).bounding_volume;
   }
 
   template <class BVH>
-  static KOKKOS_FUNCTION
-// FIXME_HIP See https://github.com/arborx/ArborX/issues/553
-#ifdef __HIP_DEVICE_COMPILE__
-      auto
-#else
-      auto const &
-#endif
-      getIndexable(BVH const &bvh, int i)
+  static KOKKOS_FUNCTION auto const &getIndexable(BVH const &bvh, int i)
   {
     return bvh._indexable_getter(getValue(bvh, i));
   }