This is not necessary anymore. Tested on frontier with rocm 5.2 and rocm 5.4.3. I don't have the input data to run DBSCAN (I keep losing it), so @aprokop should probably check that it gets the same results.

The different path for HIP was first introduced in #575.