@aprokop 2023-05-26
Looking at [this](https://github.com/kokkos/kokkos/issues/5285#issue-1320161352), Kokkos supports 4.5 for 3.7. Should this PR be delayed until we require Kokkos 4.0 (assuming the performance checks out)?

> Tested on frontier with rocm 5.2 and rocm 5.4.3.

Test how?