@Rombur 2023-05-26
> Should this PR be delayed until we require Kokkos 4.0 (assuming the performance checks out)?

I am neutral on this but I doubt many people are still using ROCm 4.5 when there are other versions available on crusher/frontier

> Tested how?

I tested on the MD problem. I increased the size of the problem to get better timing.