@aprokop 2023-05-26
How did you realize it wasn't necessary? Were there any compiler release notes, or did you just try it out?