@Rombur 2023-05-26
Damien ask me to check if that was still necessary. The timings were exactly the same so I didn't bother looking at the assembly.