@aprokop 2023-05-30
> I don't have the input data to run DBSCAN (I keep losing it), so @aprokop should probably check that it gets the same results.

I quickly tried to do that, but ran into some runtime issues with `libomp` and stuff. 
@Rombur If you have time to test it out, the data is available in the project shared directory (Orion).