@dalg24 2023-05-31
We need to look at the assembly, whether the loads are vectorized.