@Rombur 2023-05-31
> We need to look at the assembly, whether the loads are vectorized.

The ISA code is about 650,000 lines. They are differences everywhere: sometimes the register used is different, sometimes the order of the instructions is different, sometimes one version of the code has more instructions than the other...  Neither version of the code is using `flat_load_dwordx` if that was question.