@Rombur 2023-05-31
> If you have time to test it out, the data is available in the project shared directory

The two versions of the code give the same timings

Master
```
Regions:


-------------------------------------------------------------------------
Kernels:

- ArborX::Experimental::HalfTraversal
 (ParFor)   5.857340 1 5.857340 91.083567 34.494111
- ArborX::TreeConstruction::generate_hierarchy
 (ParFor)   0.229104 1 0.229104 3.562643 1.349203
- ArborX::DBSCAN::compute_cluster_indices
 (ParFor)   0.095097 1 0.095097 1.478787 0.560029
- ArborX::DBSCAN::finalize_labels
 (ParFor)   0.092402 1 0.092402 1.436881 0.544159
- ArborX::DBSCAN::compute_cluster_sizes
 (ParFor)   0.089920 1 0.089920 1.398286 0.529543
- ArborX::Algorithms::iota
 (ParFor)   0.017666 2 0.008833 0.274717 0.104038
- ArborX::TreeConstruction::calculate_bounding_box_of_the_scene
 (ParRed)   0.015498 1 0.015498 0.240998 0.091268
- ArborX::DBSCAN::compute_cluster_offset_with_filter
 (ParScan)  0.010346 1 0.010346 0.160883 0.060928
- ArborX::TreeConstruction::project_primitives_onto_space_filling_curve
 (ParFor)   0.009942 1 0.009942 0.154602 0.058549
- ArborX::DBSCAN::mark_noise
 (ParFor)   0.006532 1 0.006532 0.101574 0.038467
- Kokkos::View::initialization [ArborX::DBSCAN::cluster_sizes] via memset
 (ParFor)   0.003436 2 0.001718 0.053429 0.020234
- Kokkos::ViewFill-1D
 (ParFor)   0.002602 1 0.002602 0.040464 0.015324
- ArborX::Algorithms::exclusive_scan
 (ParScan)  0.000570 1 0.000570 0.008865 0.003357
- Kokkos::ViewCopy-1D
 (ParFor)   0.000253 1 0.000253 0.003934 0.001490
- Kokkos::View::destruction []
 (ParFor)   0.000024 2 0.000012 0.000371 0.000140

-------------------------------------------------------------------------
Summary:

Total Execution Time (incl. Kokkos + non-Kokkos):                  16.98070 seconds
Total Time in Kokkos kernels:                                       6.43073 seconds
   -> Time outside Kokkos kernels:                                 10.54996 seconds
   -> Percentage in Kokkos kernels:                                   37.87 %
Total Calls to Kokkos Kernels:                                           18

-------------------------------------------------------------------------
```

PR
```
Regions:


-------------------------------------------------------------------------
Kernels:

- ArborX::Experimental::HalfTraversal
 (ParFor)   5.850497 1 5.850497 91.072034 21.217827
- ArborX::TreeConstruction::generate_hierarchy
 (ParFor)   0.229177 1 0.229177 3.567498 0.831150
- ArborX::DBSCAN::compute_cluster_indices
 (ParFor)   0.095010 1 0.095010 1.478975 0.344569
- ArborX::DBSCAN::finalize_labels
 (ParFor)   0.092320 1 0.092320 1.437103 0.334814
- ArborX::DBSCAN::compute_cluster_sizes
 (ParFor)   0.089609 1 0.089609 1.394905 0.324983
- ArborX::Algorithms::iota
 (ParFor)   0.017979 2 0.008990 0.279877 0.065205
- ArborX::TreeConstruction::calculate_bounding_box_of_the_scene
 (ParRed)   0.015806 1 0.015806 0.246048 0.057324
- ArborX::DBSCAN::compute_cluster_offset_with_filter
 (ParScan)  0.010323 1 0.010323 0.160694 0.037438
- ArborX::TreeConstruction::project_primitives_onto_space_filling_curve
 (ParFor)   0.009911 1 0.009911 0.154281 0.035944
- ArborX::DBSCAN::mark_noise
 (ParFor)   0.006511 1 0.006511 0.101353 0.023613
- Kokkos::View::initialization [ArborX::DBSCAN::cluster_sizes] via memset
 (ParFor)   0.003436 2 0.001718 0.053484 0.012461
- Kokkos::ViewFill-1D
 (ParFor)   0.002606 1 0.002606 0.040565 0.009451
- ArborX::Algorithms::exclusive_scan
 (ParScan)  0.000569 1 0.000569 0.008855 0.002063
- Kokkos::ViewCopy-1D
 (ParFor)   0.000253 1 0.000253 0.003938 0.000917
- Kokkos::View::destruction []
 (ParFor)   0.000025 2 0.000013 0.000390 0.000091

-------------------------------------------------------------------------
Summary:

Total Execution Time (incl. Kokkos + non-Kokkos):                  27.57350 seconds
Total Time in Kokkos kernels:                                       6.42403 seconds
   -> Time outside Kokkos kernels:                                 21.14946 seconds
   -> Percentage in Kokkos kernels:                                   23.30 %
Total Calls to Kokkos Kernels:                                           18

-------------------------------------------------------------------------
```

The PR code is slower because the non-kokkos part is slower. I didn't touch that part so the timing difference is unrelated to the PR.